Crate fem_2d

A 2D FEM implementation which supports aniosotropic hp-refinements

This Library is primarily focused on Solving the Maxwell Eigenvalue Problem on a PEC-terminated Mesh; however, its functionality can be easily extended by creating custom implementations of the Integral and ShapeFn Traits

Example

This example shows how to solve the Maxwell Eigenvalue Problem on standard waveguide and print the electric fields to a VTK file

use fem_2d::prelude::*;

// Load a standard air-filled waveguide mesh from a JSON file
let mut mesh = Mesh::from_file("./test_input/test_mesh_a.json").unwrap();

// Set the polynomial expansion order to 3 in both directions on all Elems
mesh.set_global_expansion_orders([3, 3]);

// Anisotropically refine all Elems along the x-direction
mesh.h_refine_elems(vec![0, 1, 2, 3], HRef::U(None));

// Construct a domain with Dirichlet boundary conditions
let domain = Domain::new(mesh);
println!("Constructed Domain with {} DoFs", domain.dofs.len());

// Construct a generalized eigenvalue problem for the Electric Field
    // (in parallel using the Rayon Global ThreadPool)
let gep = domain.galerkin_sample_gep_parallel::<KOLShapeFn, CurlCurl, L2Inner>(None);

// Solve the generalized eigenvalue problem using Nalgebra's Eigen-Decomposition
    // look for an eigenvalue close to 2.0
let solution = nalgebra_solve_gep(gep, 2.0).unwrap();
println!("Found Eigenvalue: {:.15}", solution.value);

// Construct a solution-field-space over the Domain with 64 samples on each "shell" Elem
let mut field_space = UniformFieldSpace::new(&domain, [8, 8]);

// Compute the Electric Field in the X- and Y-directions (using the same ShapeFns as above)
let e_field_names = field_space.xy_fields::<KOLShapeFn>("E", solution.normalized_eigenvector()).unwrap();

// Compute the magnitude of the Electric Field
field_space.expression_2arg(e_field_names, "E_mag", |ex, ey| (ex.powi(2) + ey.powi(2)).sqrt()).unwrap();

// Print "E_x", "E_y" and "E_mag" to a VTK file
field_space.print_all_to_vtk("./test_output/electric_field_solution.vtk").unwrap();

Modules

basis

Structures and Traits for Basis Function Evaluation

domain

Structures to define the geometric characteristics and refinement state of an FEM Domain

integration

Structures and functions to assist in the integration of Basis Functions

linalg

Structures and functions to solve Generalized Eigenvalue Problems

prelude

Convenient Re-Exports