Struct fem_2d::domain::fields::UniformFieldSpace

pub struct UniformFieldSpace<'d> { /* fields omitted */ }

A collection of Field Solutions over a Domain.

Solutions can be operated on and printed to VTK files for visualization.

Implementations

impl<'d> UniformFieldSpace<'d>

pub fn new(domain: &'d Domain, densities: [usize; 2]) -> Self

Generate a FieldSpace over a Domain

domain: is a reference to a Domain which must outlive this Structure. It is used for all subsequent computations

densities: the number of evaluation points in the u and v directions respectively. These values apply to Elems on the shell of the Domain (their ancestor Elems will have more evaluation points)

pub fn xy_fields<SF: ShapeFn>(
&mut self,
vector_name: &'static str,
eigenvector: Vec<f64>
) -> Result<[String; 2], String>

Use an eigenvector and associated ShapeFn to compute the X and Y fields over the Domain

The X and Y field quantities will be stored as X_{vector_name} and Y_{vector_name} respectively. The Names are returned in an array in that order.

pub fn print_all_to_vtk(&self, path: impl AsRef<str>) -> Result<()>

create a VTK file at the designated path (with the file name.vtk) including all Field Quantities

These files can be plotted using Visit

pub fn print_quantities_to_vkt(
&self,
path: impl AsRef<str>,
quantity_names: Vec<String>
) -> Result<()>

create a VTK file at the designated path (with the file name.vtk) including a list of Field Quantities

These files can be plotted using Visit

pub fn map_to_quantity<F>(
&mut self,
name: impl AsRef<str>,
result_name: impl AsRef<str>,
operator: F
) -> Result<(), String> where
F: Fn(&f64) -> f64 + Copy,

map an operation over a field quantity (name) and store the result in a new quantity (result_name)

pub fn expression_2arg<F>(
&mut self,
operand_names: [impl AsRef<str>; 2],
result_name: impl AsRef<str>,
expression: F
) -> Result<(), String> where
F: Fn(f64, f64) -> f64,

evaluate an expression of two field quantities and store the result in a new quantity (result_name)