Module fem_2d::domain::mesh

The internal geometric structure of a Domain

Modules

edge

elem

element

h_refinement

node

p_refinement

space

Structs

Mesh

Information used to Define the geometric structure and refinement state of a Domain.

Constants

EXPECTED_NUM_H_REFINEMENTS

The expected “h-Refinement” depth. This determines the stack allocation size of some SmallVecs related to h-Refinement

MAX_POLYNOMIAL_ORDER

Maximum Polynomial expansion. p-Refinements will fail when Elem’s expansion orders exceed this value.

MIN_EDGE_LENGTH

Minimum Edge length in parametric space. h-Refinements will fail after edges are smaller than this value.