Struct fem_2d::domain::mesh::Mesh

pub struct Mesh {
    pub elements: Vec<Arc<Element>>,
    pub elems: Vec<Elem>,
    pub nodes: Vec<Node>,
    pub edges: Vec<Edge>,
}

Information used to Define the geometric structure and refinement state of a Domain.

Fields
elements: Vec<Arc<Element>>elems: Vec<Elem>nodes: Vec<Node>edges: Vec<Edge>
Implementations

impl Mesh

pub fn blank() -> Self

Construct a completely empty Mesh

pub fn from_file(path: impl AsRef<str>) -> Result<Self>

Construct a Mesh from a JSON file with the following format

The first “Element” and “Node” describe the meaning of each variable

The following entries in each array describe this two element mesh:

    3               4               5
0.5 *---------------*---------------*
    |               |               |
    |      air      |    teflon     |
    |               |               |
0.0 *---------------*---------------*
 y  0               1               2
 x 0.0             1.0             2.0

mesh.json

{
    "Elements": [
        {
            "materials": [eps_rel_re, eps_rel_im, mu_rel_re, mu_rel_im],
            "node_ids": [node_0_id, node_1_id, node_2_id, node_3_id],
        },
        {
            "materials": [1.0, 0.0, 1.0, 0.0],
            "node_ids": [1, 2, 4, 5],
        },
        {
            "materials": [1.2, 0.0, 0.9999, 0.0],
            "node_ids": [2, 3, 5, 6],
        }
    ],
    "Nodes": [
        [x_coordinate, y_coordinate],
        [0.0, 0.0],
        [1.0, 0.0],
        [2.0, 0.0],
        [0.0, 0.5],
        [1.0, 0.5],
        [2.0, 0.5],
    ]
}

pub fn export_to_json(&self, path: impl AsRef<str>) -> Result<()>

Print the mesh to a JSON file specified by path.

pub fn elem_points(&self, elem_id: usize) -> [&Point; 4]

Get the four Points composing an Elem

pub fn elem_diag_points(&self, elem_id: usize) -> [&Point; 2]

Get the “smallest” and “largest” Points composing an Elem

pub fn edge_points(&self, edge_id: usize) -> [&Point; 2]

Get the two Points composing an Edge

pub fn descendant_elems(
&self,
elem_id: usize,
include_starting_elem: bool
) -> Result<Vec<usize>, String>

Get a list of an Elems descendant’s IDs

pub fn ancestor_elems(
&self,
elem_id: usize,
include_starting_elem: bool
) -> Result<Vec<usize>, String>

Get a list of an Elems ancestors’s IDs

pub fn descendant_edges(
&self,
edge_id: usize,
include_starting_edge: bool
) -> Result<Vec<usize>, String>

Get a list of an Edges descendant’s IDs

pub fn max_expansion_orders(&self) -> [u8; 2]

pub fn global_h_refinement(&mut self, refinement: HRef) -> Result<(), HRefError>

Apply an HRef to all Elems that have not been h-refined

pub fn h_refine_elems(
&mut self,
elems: Vec<usize>,
refinement: HRef
) -> Result<(), HRefError>

Apply an HRef to a list of Elems by their ID

pub fn h_refine_with_filter<F>(&mut self, filt: F) -> Result<(), HRefError> where
F: Fn(&Elem) -> Option<HRef>,

h-refine Elems according to an external filter function

pub fn execute_h_refinements(
&mut self,
refinements: Vec<(usize, HRef)>
) -> Result<(), HRefError>

Execute a series of HRefs on Elems specified by their id

pub fn global_p_refinement(&mut self, refinement: PRef) -> Result<(), PRefError>

Apply a PRef to all Elems

pub fn p_refine_elems(
&mut self,
elems: Vec<usize>,
refinement: PRef
) -> Result<(), PRefError>

Apply a PRef to a list of Elems by their ID

pub fn p_refine_with_filter<F>(&mut self, filt: F) -> Result<(), PRefError> where
F: Fn(&Elem) -> Option<PRef>,

p-refine Elems according to an external filter function

pub fn execute_p_refinements(
&mut self,
refinements: Vec<(usize, PRef)>
) -> Result<(), PRefError>

Execute a series of PRefs on Elems specified by their id

pub fn set_global_expansion_orders(
&mut self,
orders: [u8; 2]
) -> Result<(), PRefError>

Set the expansion orders on all Elems

pub fn set_expansion_on_elems(
&mut self,
elems: Vec<usize>,
orders: [u8; 2]
) -> Result<(), PRefError>

Set the expansion orders on a list of Elems by their ID

pub fn set_expansions_with_filter<F>(
&mut self,
filt: F
) -> Result<(), PRefError> where
F: Fn(&Elem) -> Option<[u8; 2]>,

set expansion orders on Elems according to an external filter function

pub fn set_expansion_orders(
&mut self,
poly_orders: Vec<(usize, [u8; 2])>
) -> Result<(), PRefError>