Struct fem_2d::domain::Domain

pub struct Domain {
    pub mesh: Mesh,
    pub dofs: Vec<DoF>,
    pub basis_specs: Vec<Vec<BasisSpec>>,
}

High Level Description of an FEM Domain

Fields
mesh: Meshdofs: Vec<DoF>

Degrees of Freedom (collections of BasisSpecs matched via Tangential Continuity)

basis_specs: Vec<Vec<BasisSpec>>

Individual BasisSpecs sorted by their associated Elem

Implementations

impl Domain

pub fn blank() -> Self

pub fn from_mesh(mesh: Mesh) -> Self

pub fn elems<'a>(&'a self) -> impl Iterator<Item = &'a Elem> + '_

Iterate over all Elems in the mesh

pub fn edges<'a>(&'a self) -> impl Iterator<Item = &'a Edge> + '_

Iterate over all Edges in the mesh

pub fn nodes<'a>(&'a self) -> impl Iterator<Item = &'a Node> + '_

Iterate over all Nodes in the mesh

pub fn local_basis_specs(
&self,
elem_id: usize
) -> Result<&Vec<BasisSpec>, String>

Retrieve a list of BasisSpecs on an Elem by ID

pub fn descendant_basis_specs(
&self,
elem_id: usize
) -> Result<Vec<(usize, &Vec<BasisSpec>)>, String>

Retrieve a list of an Elems descendant BasisSpecs (All the BasisSpecs on its descendant Elems)

pub fn ancestor_basis_specs(
&self,
elem_id: usize
) -> Result<Vec<&BasisSpec>, String>

Retrieve a list of an Elems ancestor BasisSpecs (All the BasisSpecs on its ancestor Elems)

pub fn galerkin_sample_gep<SF, AI, BI>(
&self,
num_gauss_quad: Option<usize>
) -> GEP where
SF: ShapeFn,
AI: Integral,
BI: Integral,

Fill two system matrices using the domains it’s Basis Space as the Testing Space. Return a Generalized Eigenproblem (GEP)

All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs

Two Integrals: AI and BI must be specified. These are used to populate the A and B matrices respectively

A ShapeFn SF must also be specified. This is used to evaluate the Domains BasisSpecs

pub fn galerkin_sample_gep_parallel<SF, AI, BI>(
&self,
num_gauss_quad: Option<usize>
) -> GEP where
SF: ShapeFn,
AI: Integral,
BI: Integral,

Same as galerkin_sample_gep, except matrices are filled in parallel using the Rayon Global ThreadPool

pub fn get_basis_spec(&self, address: BSAddress) -> Result<&BasisSpec, String>

Retrieve a BasisSpec at a particular BSAddress

Returns an error if the designated Elem does not exist, or does not have that BasisSpec