Trait fem_2d::integration::Integral

pub trait Integral: Sync + Send {
    fn with_weights(u_weights: &[f64], v_weights: &[f64]) -> Self;

    fn integrate<SF: ShapeFn>(
        &self, 
        p_dir: BasisDir, 
        q_dir: BasisDir, 
        p_orders: [usize; 2], 
        q_orders: [usize; 2], 
        p_basis: &BasisFn<SF>, 
        q_basis: &BasisFn<SF>
    ) -> IntegralResult;

    fn integrate_by_parts<SF: ShapeFn>(
        &self, 
        p_dir: BasisDir, 
        q_dir: BasisDir, 
        p_orders: [usize; 2], 
        q_orders: [usize; 2], 
        p_basis: &BasisFn<SF>, 
        q_basis: &BasisFn<SF>
    ) -> IntegralResult;
}

A trait to describe an “integrator” which can compute 2D integrals over some function of two BasisFn’s

Required methods

fn with_weights(u_weights: &[f64], v_weights: &[f64]) -> Self

Construct the Integral with u and v directed Gauss-Leg-Quad weights.

The weight vectors must match the dimension of the BasisFns used in later calls to integrate or integrate_by_parts

fn integrate<SF: ShapeFn>(
&self,
p_dir: BasisDir,
q_dir: BasisDir,
p_orders: [usize; 2],
q_orders: [usize; 2],
p_basis: &BasisFn<SF>,
q_basis: &BasisFn<SF>
) -> IntegralResult

Compute an integral between BasisFn’s P and Q, where P and Q both have a direction (BasisDir) and orders i and j.

fn integrate_by_parts<SF: ShapeFn>(
&self,
p_dir: BasisDir,
q_dir: BasisDir,
p_orders: [usize; 2],
q_orders: [usize; 2],
p_basis: &BasisFn<SF>,
q_basis: &BasisFn<SF>
) -> IntegralResult

Compute an integral-by-parts between BasisFn’s P and Q, where P and Q both have a direction (BasisDir) and orders i and j.

This function may still return a the Full variant of IntegralResult if the solution is known to be zero along the edges.