Trait fem_2d::Integral

pub trait Integral {
    fn with_weights(u_weights: &[f64], v_weights: &[f64]) -> Self;

    fn integrate<SF>(
        &self, 
        p_dir: ParaDir, 
        q_dir: ParaDir, 
        p_orders: [usize; 2], 
        q_orders: [usize; 2], 
        p_basis: &BasisFn<SF>, 
        q_basis: &BasisFn<SF>
    ) -> IntegralResult
    where
        SF: ShapeFn;

    fn integrate_by_parts<SF>(
        &self, 
        p_dir: ParaDir, 
        q_dir: ParaDir, 
        p_orders: [usize; 2], 
        q_orders: [usize; 2], 
        p_basis: &BasisFn<SF>, 
        q_basis: &BasisFn<SF>
    ) -> IntegralResult
    where
        SF: ShapeFn;
}

A trait to describe an “integrator” which can compute 2D integrals over some function of two BasisFn’s

Required methods

fn with_weights(u_weights: &[f64], v_weights: &[f64]) -> Self

fn integrate<SF>(
&self,
p_dir: ParaDir,
q_dir: ParaDir,
p_orders: [usize; 2],
q_orders: [usize; 2],
p_basis: &BasisFn<SF>,
q_basis: &BasisFn<SF>
) -> IntegralResult where
SF: ShapeFn,

Compute an integral between BasisFn’s P and Q, where P and Q both have a parametric direction (ParaDir) and orders i and j.

fn integrate_by_parts<SF>(
&self,
p_dir: ParaDir,
q_dir: ParaDir,
p_orders: [usize; 2],
q_orders: [usize; 2],
p_basis: &BasisFn<SF>,
q_basis: &BasisFn<SF>
) -> IntegralResult where
SF: ShapeFn,

Compute an integral-by-parts between BasisFn’s P and Q, where P and Q both have a parametric direction (ParaDir) and orders i and j. This function may still return a “Full” integral result if the solution is known to be zero along the edges.