Crate fem_2d

Structs

BasisFn

Structure used to evaluate ShapeFn’s and their derivatives over some area

BasisFnSampler

Structure used to generate BasisFn’s over Elem’s in a Domain. This structure contains settings for a

CurlProduct

Elem

KOLShapeFn

A simple Hierarchical type Shape Function

L2InnerProduct

M2D

MaxOrthoShapeFn

An advanced Hierarchical Type Shape Function which maximizes orthogonality between polynomial orders.

Point

V2D

Enums

IntegralResult

Return type of an Integral

Traits

Integral

A trait to describe an “integrator” which can compute 2D integrals over some function of two BasisFn’s

ShapeFn

Hierarchical Shape Function along a single direction (defined over (-1.0, +1.0)). KOLShapeFn and MaxOrthoShapeFn implement this trait. Alternate Hierarchical Basis Functions can be used by implementing this trait.